Reestrcturar los archivos de codigo con una arquitectura MVC, sin asegurar aun la completa funcionalidad del programa
