+++ Interfaz/src/Arboles/Arboles.java
@@ -2,8 +2,8 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package Arboles;
-import newpackage.Usuario;
+package arboles;
+import model.Usuario;
 /**
  *
  * @author pablo
