+++ Interfaz/src/Main.java
@@ -2,9 +2,11 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/Classes/Class.java to edit this template
  */
-package newpackage;
 
-import Arboles.*;
+
+import arboles.*;
+import model.Usuario;
+import view.Pantalla;
 
 
 /**
