+++ Interfaz/src/view/EliminarLibro.java
@@ -2,21 +2,21 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
  */
-package newpackage;
+package view;
 
-import Arboles.Arboles;
+import arboles.*;
 import javax.swing.JOptionPane;
 
 /**
  *
  * @author pablo
  */
-public class DevolverLibro extends javax.swing.JFrame {
+public class EliminarLibro extends javax.swing.JFrame {
 
     /**
-     * Creates new form DevolverLibro
+     * Creates new form EliminarLibro
      */
-    public DevolverLibro() {
+    public EliminarLibro() {
         initComponents();
     }
 
@@ -40,7 +40,7 @@ public class DevolverLibro extends javax.swing.JFrame {
 
         setDefaultCloseOperation(javax.swing.WindowConstants.EXIT_ON_CLOSE);
 
-        jButton1.setText("Atras");
+        jButton1.setText("Salir");
         jButton1.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jButton1ActionPerformed(evt);
@@ -54,9 +54,9 @@ public class DevolverLibro extends javax.swing.JFrame {
         });
 
         jLabel2.setFont(new java.awt.Font("Segoe UI", 1, 18)); // NOI18N
-        jLabel2.setText("Devolver libro");
+        jLabel2.setText("Eliminar libro");
 
-        jButton2.setText("Devolver");
+        jButton2.setText("Eliminar");
         jButton2.addActionListener(new java.awt.event.ActionListener() {
             public void actionPerformed(java.awt.event.ActionEvent evt) {
                 jButton2ActionPerformed(evt);
@@ -68,7 +68,7 @@ public class DevolverLibro extends javax.swing.JFrame {
         jLabel1.setToolTipText("");
 
         jLabel6.setFont(new java.awt.Font("Segoe UI", 3, 14)); // NOI18N
-        jLabel6.setText("Ingresa el libro que desea devolver");
+        jLabel6.setText("Ingresa el libro que desea eliminar");
 
         javax.swing.GroupLayout jPanel2Layout = new javax.swing.GroupLayout(jPanel2);
         jPanel2.setLayout(jPanel2Layout);
@@ -87,16 +87,16 @@ public class DevolverLibro extends javax.swing.JFrame {
                             .addComponent(jButton2)
                             .addComponent(jLabel6, javax.swing.GroupLayout.PREFERRED_SIZE, 310, javax.swing.GroupLayout.PREFERRED_SIZE)))
                     .addGroup(jPanel2Layout.createSequentialGroup()
-                        .addGap(129, 129, 129)
+                        .addGap(117, 117, 117)
                         .addComponent(jLabel2)))
                 .addContainerGap(15, Short.MAX_VALUE))
         );
         jPanel2Layout.setVerticalGroup(
             jPanel2Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
             .addGroup(jPanel2Layout.createSequentialGroup()
-                .addGap(12, 12, 12)
+                .addContainerGap()
                 .addComponent(jLabel2, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.UNRELATED)
+                .addGap(18, 18, 18)
                 .addComponent(jLabel6, javax.swing.GroupLayout.PREFERRED_SIZE, 28, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(81, 81, 81)
                 .addComponent(jLabel1)
@@ -104,7 +104,7 @@ public class DevolverLibro extends javax.swing.JFrame {
                 .addComponent(txtIdLibro, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
                 .addGap(18, 18, 18)
                 .addComponent(jButton2)
-                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 114, Short.MAX_VALUE)
+                .addPreferredGap(javax.swing.LayoutStyle.ComponentPlacement.RELATED, 112, Short.MAX_VALUE)
                 .addComponent(jButton1)
                 .addGap(37, 37, 37))
         );
@@ -117,7 +117,9 @@ public class DevolverLibro extends javax.swing.JFrame {
         );
         jPanel1Layout.setVerticalGroup(
             jPanel1Layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPanel2, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addGroup(jPanel1Layout.createSequentialGroup()
+                .addComponent(jPanel2, javax.swing.GroupLayout.PREFERRED_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.PREFERRED_SIZE)
+                .addGap(0, 0, Short.MAX_VALUE))
         );
 
         javax.swing.GroupLayout layout = new javax.swing.GroupLayout(getContentPane());
@@ -128,19 +130,18 @@ public class DevolverLibro extends javax.swing.JFrame {
         );
         layout.setVerticalGroup(
             layout.createParallelGroup(javax.swing.GroupLayout.Alignment.LEADING)
-            .addComponent(jPanel1, javax.swing.GroupLayout.Alignment.TRAILING, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
+            .addComponent(jPanel1, javax.swing.GroupLayout.DEFAULT_SIZE, javax.swing.GroupLayout.DEFAULT_SIZE, Short.MAX_VALUE)
         );
 
         pack();
     }// </editor-fold>//GEN-END:initComponents
 
     private void jButton1ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton1ActionPerformed
-        // TODO add your handling code here:        // TODO add your handling code here:
-
-        Panta1 pantausu = new Panta1();
-        pantausu.setVisible(true);
-        pantausu.setResizable(false);
-        pantausu.setLocationRelativeTo(null);
+        // TODO add your handling code here:     
+        PantallaAdmin ingre = new PantallaAdmin();
+        ingre.setVisible(true);
+        ingre.setResizable(false);
+        ingre.setLocationRelativeTo(null);
         this.setVisible(false);
     }//GEN-LAST:event_jButton1ActionPerformed
 
@@ -150,33 +151,37 @@ public class DevolverLibro extends javax.swing.JFrame {
 
     private void jButton2ActionPerformed(java.awt.event.ActionEvent evt) {//GEN-FIRST:event_jButton2ActionPerformed
         // TODO add your handling code here:
-        
-        int id = Integer.parseInt(txtIdLibro.getText());
 
-        Libro libro = Arboles.libros.buscarLibroPorID(id);
-        if (libro != null && libro.getEstado() != true) {
-            
-            Arboles.libros.cambiarEstadoLibroPorID(id);
-            Arboles.usuarioActual.retirarLibroPrestado(id);
-            JOptionPane.showInternalMessageDialog(null, "El libro se ha devuelto, gracias :)");
-            
-        }else{
-            JOptionPane.showInternalMessageDialog(null, "El libro no existe");
+        String idString = txtIdLibro.getText();
+        
+        int id;
+
+        try {
+            id = Integer.parseInt(idString);
+        } catch (NumberFormatException e) {
+            JOptionPane.showMessageDialog(this, "No existe el ID indicado", "Error", JOptionPane.ERROR_MESSAGE);
+            return;
         }
-        
-        
-     /*   int id = Integer.parseInt(txtIdLibro.getText());
-    ControladorLibro.devolverLibro(id);
-    JOptionPane.showMessageDialog(this, "El libro ha sido devuelto con éxito.", "Libro devuelto", JOptionPane.INFORMATION_MESSAGE);
+
+        Arboles.libros.eliminarLibro(id);
+        JOptionPane.showMessageDialog(null, "Libro eliminado de forma exitosa");
+        txtIdLibro.setText("");
+        txtIdLibro.requestFocus();
+
+   
+        /*  int id = Integer.parseInt(txtIdLibro.getText());
+
+    ControladorLibro.eliminarLibro(id);
+
+    JOptionPane.showMessageDialog(this, "El libro fue eliminado con éxito.", "Libro eliminado", JOptionPane.INFORMATION_MESSAGE);
+
     txtIdLibro.setText("");
     txtIdLibro.requestFocus();*/
-
     }//GEN-LAST:event_jButton2ActionPerformed
 
     /**
      * @param args the command line arguments
      */
-
 
     // Variables declaration - do not modify//GEN-BEGIN:variables
     private javax.swing.JButton jButton1;
@@ -188,4 +193,5 @@ public class DevolverLibro extends javax.swing.JFrame {
     private javax.swing.JPanel jPanel2;
     private javax.swing.JTextField txtIdLibro;
     // End of variables declaration//GEN-END:variables
+
 }
