+++ Interfaz/src/view/Registro.form
