+++ Interfaz/src/view/Registro.java
@@ -2,10 +2,11 @@
  * Click nbfs://nbhost/SystemFileSystem/Templates/Licenses/license-default.txt to change this license
  * Click nbfs://nbhost/SystemFileSystem/Templates/GUIForms/JFrame.java to edit this template
  */
-package newpackage;
+package view;
 
-import Arboles.Arboles;
-import java.util.ArrayList;
+import arboles.*;
+import model.Usuario;
+
 import javax.swing.JOptionPane;
 
 /**
